Create factory for *CommandBarComponentProp
